refactor: reorganize modules into subpackages
